+++ HISTORY.rst
@@ -2,6 +2,10 @@
 History
 =======
 
+* Add Django 4.0+ fixer to remove ``USE_L10N = True`` setting.
+
+  Thanks to Johnny Metz in `PR #173 <https://github.com/adamchainz/django-upgrade/pull/173>`__.
+
 1.8.1 (2022-08-25)
 ------------------